+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Control Panel: updated to work with OpenJDK;
 - Auto-updater: cleanup previous backup folder before updating;
 
+### Fixed
+- Relevant multiple attribute column not shown in tables;
+
 ## [4.0.14] - 2021-01-28
 ### Fixed
 - Error generating polygon (geometry) in Collect Mobile