+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.0.15] - NOT RELEASED YET
+## [4.0.16] - NOT RELEASED YET
+### Added
+- Allow usage of calculated key attributes with "only one time" option;
+
+## [4.0.15] - 2021-02-22
 ### Added
 - Control Panel: updated to work with OpenJDK;
 - Auto-updater: cleanup previous backup folder before updating;